refactor(library-notes): move the refusal copy out of the controller (review findings 1, 8)
The copy table, both helpers and the loguru import now live in
Library/library_notes_lasting_sync_state.py with the rest of this screen's
presentation state. Both fallback strings moved into check_failure_line,
keyed on root_id == "" => setup, so each Check path spends one line on it.

The controller's whole diff against dev is now five lines: one import and
two call sites. Measured 2024; the pin is restored from the rejected 2128 to
that exact measurement (the +1 is the import line -- 2023 would fail the
ceiling test). Ratchet FAILED-name set is identical to dev again.

Inventory re-pinned: the log statement moved files, same digest, text
unchanged. Canaries in this task's tests are now /zqleakcanary/... per
lessons-testing-evidence; the pre-existing ones are left alone.
