+++ tldw_chatbook/UI/Library_Modules/library_notes_sync_controller.py
@@ -6,8 +6,6 @@ from collections.abc import Awaitable, Callable
 from dataclasses import replace
 from typing import Never, Protocol
 
-from loguru import logger
-
 from tldw_chatbook.Library.library_notes_lasting_sync_state import (
     LASTING_SYNC_HISTORY_PAGE_SIZE,
     LastingSyncApplyBlocker,
@@ -19,6 +17,7 @@ from tldw_chatbook.Library.library_notes_lasting_sync_state import (
     LastingSyncRootRow,
     LibraryNotesLastingSyncSnapshot,
     build_reconciliation_review,
+    check_failure_line,
     initial_lasting_sync_snapshot,
     set_setup_value,
     validate_lasting_sync_history_page,
@@ -49,136 +48,6 @@ from tldw_chatbook.Notes.notes_sync_models import (
     NotesSyncDirection,
     NotesSyncOperationState,
 )
-
-
-# TASK-32243: a refused Check used to render one fixed string for every cause,
-# with nothing logged anywhere. Each bounded reason the runtime and coordinator
-# can name gets its own sentence and its own next action here; anything else
-# keeps the generic line but still leaves a typed warning behind.
-_CHECK_REFUSAL_COPY: dict[str, str] = {
-    "private_path_overlap": (
-        "That folder is inside Chatbook's own data directory. "
-        "Pick a folder outside it, then Check again."
-    ),
-    "private_path_check_failed": (
-        "Chatbook could not confirm that folder sits outside its own data. "
-        "Pick a folder in your home directory, then Check again."
-    ),
-    "file_notes_overlap": (
-        "That folder is already in use by Folder files. "
-        "Pick a different folder, then Check again."
-    ),
-    "lasting_root_overlap": (
-        "That folder is already connected. Manage it under Sync folders "
-        "instead of adding it again."
-    ),
-    "passive_process": (
-        "Another Chatbook window is using that folder. "
-        "Close it, then Check again."
-    ),
-    "lock_unavailable": (
-        "Chatbook could not claim that folder. "
-        "Restart Chatbook, then Check again."
-    ),
-    "root_lease_unavailable": (
-        "That folder could not be claimed for sync. "
-        "Close any other Chatbook window using it, then Check again."
-    ),
-    "unsupported_metadata": (
-        "Some files there use a permission model this sync can't track. "
-        "Make every file in the folder writable, then Check again."
-    ),
-    # The bare gate name, which is what a *listing* failure reports. Every
-    # per-file refusal replaces it with the dominant gate below.
-    "root_discovery_incomplete": (
-        "Some files in that folder couldn't be read. "
-        "Check that you can open everything in it, then Check again."
-    ),
-    "mixed_newlines": (
-        "Some files there use a mix of line endings. "
-        "Save them with one style, then Check again."
-    ),
-    "unsupported_newline": (
-        "Some files there use older Mac line endings. "
-        "Save them with Unix or Windows line endings, then Check again."
-    ),
-    "unsupported_encoding": (
-        "Some files there are not UTF-8 text. "
-        "Convert or move them out of the folder, then Check again."
-    ),
-    "max_file_bytes_exceeded": (
-        "Some files there are larger than 10 MB. "
-        "Move them out of the folder, then Check again."
-    ),
-    "non_regular": (
-        "Some entries there aren't ordinary files. "
-        "Move them out of the folder, then Check again."
-    ),
-    "multiple_links": (
-        "Some files there have more than one name on disk. "
-        "Leave a single copy in the folder, then Check again."
-    ),
-    "link_or_reparse": (
-        "Some entries there are links rather than files. "
-        "Move them out of the folder, then Check again."
-    ),
-    "target_changed_during_read": (
-        "Something changed those files while they were being read. "
-        "Check again once the folder is settled."
-    ),
-    "notes_sync_cutover_not_admitted": (
-        "Sync isn't available while another Chatbook instance owns this "
-        "profile. Close the other instance, then Check again."
-    ),
-    "root_offline": (
-        "That folder isn't available right now. Reconnect it, then Check again."
-    ),
-    "root_unavailable": (
-        "That folder can't be read right now. Check that it still exists and "
-        "that you can open it, then Check again."
-    ),
-    "root_not_directory": "That path isn't a folder. Pick a folder, then Check again.",
-    "root_link_or_reparse": (
-        "That path is a link. Pick the folder it points at, then Check again."
-    ),
-    "writable_filesystem_unsupported": (
-        "Chatbook can't write safely on that filesystem. "
-        "Pick a folder on a local disk, then Check again."
-    ),
-    "sync_root_not_active": (
-        "That folder is paused. Resume it, then Check again."
-    ),
-}
-
-
-def _refusal_reason(error: BaseException) -> str:
-    """Return the bounded reason code this failure names, or an empty string.
-
-    Only codes this module already knows are ever returned, so nothing an
-    exception message happens to carry -- a path, a note title -- can reach a
-    status line or a log record through here.
-    """
-
-    code = getattr(error, "reason_code", None)
-    if type(code) is str and code in _CHECK_REFUSAL_COPY:
-        return code
-    text = str(error)
-    return text if text in _CHECK_REFUSAL_COPY else ""
-
-
-def _check_failure_line(
-    error: BaseException, fallback: str, *, root_id: str = ""
-) -> str:
-    """Name one refused Check for the user and for the log."""
-
-    reason = _refusal_reason(error)
-    logger.warning(
-        "notes sync check refused; reason={} error_type={} root_id={}",
-        reason or "unclassified",
-        type(error).__name__,
-        root_id or "pending-setup",
-    )
-    return _CHECK_REFUSAL_COPY.get(reason, fallback)
 
 
 class LastingSyncRuntimePort(Protocol):
@@ -873,11 +742,7 @@ class LibraryNotesSyncController:
                     activation=activation,
                     epoch=epoch,
                 ),
-                status_line=_check_failure_line(
-                    error,
-                    "Check failed. Review root status, then Check again.",
-                    root_id=root_id,
-                ),
+                status_line=check_failure_line(error, root_id=root_id),
             )
             self._publish()
             return
@@ -945,10 +810,7 @@ class LibraryNotesSyncController:
             self._state = replace(
                 self._state,
                 phase="configure",
-                status_line=_check_failure_line(
-                    error,
-                    "Check failed. Review the folder and settings, then try again.",
-                ),
+                status_line=check_failure_line(error),
             )
             self._publish()
             return
